Refactoring in progress: Separate header files for each class created. Getting linker errors; to be resolved.
