+++ testBox2DSDL/Shape.h
@@ -1,0 +1,8 @@
+#pragma once
+class Shape
+{
+protected:
+
+public:
+    virtual void displayShape() = 0;
+};
