+++ testBox2DSDL/Source.cpp
@@ -2,170 +2,13 @@
 #include <SDL.h>
 #include <SDL_image.h>
 #include <box2d.h>
-
-const int WIDTH = 1280;
-const int HEIGHT = 640;
-const float M2P = 20;
-const float P2M = 1 / M2P;
-
-b2World* world;
-SDL_Window* window;
-SDL_Renderer* renderer;
-
-b2Body* selectedBody = nullptr;
-
-
-class Shape
-{
-protected:
-
-public:
-    virtual void displayShape() = 0;
-};
-
-class Square : public Shape
-{
-private:
-
-public:
-
-    b2Body* addRect(int x, int y, int w, int h, bool dyn = true)
-    {
-        b2BodyDef bodydef;
-        bodydef.position.Set(x * P2M, y * P2M);
-        if (dyn)
-            bodydef.type = b2_dynamicBody;
-        b2Body* body = world->CreateBody(&bodydef);
-
-        b2PolygonShape shape;
-        shape.SetAsBox(P2M * w / 2, P2M * h / 2);
-
-        b2FixtureDef fixturedef;
-        fixturedef.shape = &shape;
-        fixturedef.density = 1.0;
-        fixturedef.restitution = 0.0;
-        body->CreateFixture(&fixturedef);
-        return body;
-    }
-
-    void displayShape()
-    {}
-
-};
-
-
-class Triangle : public Shape
-{
-public:
-
-    b2Body* addTriangle(int x, int y, int sideLength, bool dyn = true)
-    {
-        b2BodyDef bodydef;
-        bodydef.position.Set(x * P2M, y * P2M);
-        if (dyn)
-            bodydef.type = b2_dynamicBody;
-        b2Body* body = world->CreateBody(&bodydef);
-
-        b2Vec2 vertices[3];
-        float height = sideLength * sqrt(3.0f) / 2.0f;
-        vertices[0].Set(-sideLength / 2.0f, -height / 2.0f);
-        vertices[1].Set(0.0f, height / 2.0f);
-        vertices[2].Set(sideLength / 2.0f, -height / 2.0f);
-
-        b2PolygonShape shape;
-        shape.Set(vertices, 3);
-
-        b2FixtureDef fixturedef;
-        fixturedef.shape = &shape;
-        fixturedef.density = 1.0;
-        fixturedef.restitution = 1.0;
-        body->CreateFixture(&fixturedef);
-        return body;
-    }
-
-    void displayShape()
-    {}
-
-
-};
-
-class Circle : public Shape
-{
-public:
-
-    b2Body* addCircle(int x, int y, int r, bool dyn = true)
-    {
-        b2BodyDef bodydef;
-        bodydef.position.Set(x * P2M, y * P2M);
-        if (dyn)
-            bodydef.type = b2_dynamicBody;
-        b2Body* body = world->CreateBody(&bodydef);
-
-        b2CircleShape shape;
-        shape.m_radius = r * P2M;
-        shape.m_p.Set(0, 0);
-
-        b2FixtureDef fixturedef;
-        fixturedef.shape = &shape;
-        fixturedef.density = 0.5;
-        fixturedef.friction = 0.8;
-        fixturedef.restitution = 1.0;
-        body->CreateFixture(&fixturedef);
-        return body;
-    }
-
-    void displayShape()
-    {}
-
-};
-
-class Car : public Square, public Circle {
-
-private:
-
-    Square* sq1;
-    Circle* c1;
-
-public:
-
-    b2Body* carBody;
-    b2Body* wheel1;
-    b2Body* wheel2;
-    b2Body* wheel3;
-    b2Body* wheel4;
-    b2RevoluteJoint* joint1;
-    b2RevoluteJoint* joint2;
-    b2RevoluteJoint* joint3;
-    b2RevoluteJoint* joint4;
-
-    void addCar(int x, int y, int w, int h)
-
-    {
-        // Add the car body
-        carBody = sq1->addRect(x, y, w, h);
-
-        wheel1 = c1->addCircle(x - w / 2, y + h / 2, h / 2.5);
-        wheel2 = c1->addCircle(x + w / 2, y + h / 2, h / 2.5);
-        //wheel3 = addCircle(x - w / 2, y - h / 2, h / 2.5); 
-        //wheel4 = addCircle(x + w / 2, y - h / 2, h / 2.5);
-
-        // Add the revolute joints
-        b2RevoluteJointDef jointdef1, jointdef2, jointdef3, jointdef4;
-        jointdef1.Initialize(carBody, wheel1, wheel1->GetWorldCenter());
-        jointdef2.Initialize(carBody, wheel2, wheel2->GetWorldCenter());
-        //jointdef3.Initialize(carBody, wheel3, wheel3->GetWorldCenter()); 
-        //jointdef4.Initialize(carBody, wheel4, wheel4->GetWorldCenter()); 
-        jointdef1.enableMotor = true;
-        jointdef2.enableMotor = true;
-        //jointdef3.enableMotor = true;
-        //jointdef4.enableMotor = true;
-        joint1 = (b2RevoluteJoint*)world->CreateJoint(&jointdef1);
-        joint2 = (b2RevoluteJoint*)world->CreateJoint(&jointdef2);
-        //joint3 = (b2RevoluteJoint*)world->CreateJoint(&jointdef3); 
-        //joint4 = (b2RevoluteJoint*)world->CreateJoint(&jointdef4); 
-    }
-};
-
+#include "InitialSetup.h"
+#include "Constants.h"
+#include "Shape.h"
+#include "Square.h"
+#include "Triangle.h"
+#include "Circle.h"
+#include "Car.h"
 
 void drawRect(b2Vec2 vertices[], b2Vec2 center)
 {
