+++ testBox2DSDL/Square.cpp
@@ -1,0 +1,24 @@
+#include "Square.h"
+#include "InitialSetup.h"
+#include "Constants.h"
+
+b2Body* Square::addRect(int x, int y, int w, int h, bool dyn)
+{
+    b2BodyDef bodydef;
+    bodydef.position.Set(x * P2M, y * P2M);
+    if (dyn)
+        bodydef.type = b2_dynamicBody;
+    b2Body* body = world->CreateBody(&bodydef);
+
+    b2PolygonShape shape;
+    shape.SetAsBox(P2M * w / 2, P2M * h / 2);
+
+    b2FixtureDef fixturedef;
+    fixturedef.shape = &shape;
+    fixturedef.density = 1.0;
+    fixturedef.restitution = 0.0;
+    body->CreateFixture(&fixturedef);
+    return body;
+}
+
+void Square::displayShape() {}
