+++ testBox2DSDL/Triangle.cpp
@@ -1,0 +1,30 @@
+#include "Triangle.h"
+#include "InitialSetup.h"
+#include "Constants.h"
+
+b2Body* Triangle::addTriangle(int x, int y, int sideLength, bool dyn)
+{
+    b2BodyDef bodydef;
+    bodydef.position.Set(x * P2M, y * P2M);
+    if (dyn)
+        bodydef.type = b2_dynamicBody;
+    b2Body* body = world->CreateBody(&bodydef);
+
+    b2Vec2 vertices[3];
+    float height = sideLength * sqrt(3.0f) / 2.0f;
+    vertices[0].Set(-sideLength / 2.0f, -height / 2.0f);
+    vertices[1].Set(0.0f, height / 2.0f);
+    vertices[2].Set(sideLength / 2.0f, -height / 2.0f);
+
+    b2PolygonShape shape;
+    shape.Set(vertices, 3);
+
+    b2FixtureDef fixturedef;
+    fixturedef.shape = &shape;
+    fixturedef.density = 1.0;
+    fixturedef.restitution = 1.0;
+    body->CreateFixture(&fixturedef);
+    return body;
+}
+
+void Triangle::displayShape() {}
